Throw on YAML parse errors
Also refactor our YAML parsing code to use a helper.
